Initial support for "type": "module"
The package.json `type` field determines whether a `.js` file is
interpreted as CommonJS or ESM by Node.js, for version > 12.17.0.
Node.js will not allow `require()` imports of ES modules, so we must
provide a mechanism for switching between `require()` and `import()`
based on the nearest package.json type.